1.3.4 fix languageId check + change type detection
